fixed type deduction in all the std::accumulate calls

+++ src/backends/lmdb.cpp
@@ -259,12 +259,12 @@ class LMDBDatabase : public DocumentStoreMixin<DatabaseInterface> {
 
         size_t total_ksizes = std::accumulate(ksizes.data,
                                               ksizes.data + ksizes.size,
-                                              0);
+                                              (size_t)0);
         if(total_ksizes > keys.size) return Status::InvalidArg;
 
         size_t total_vsizes = std::accumulate(vsizes.data,
                                               vsizes.data + vsizes.size,
-                                              0);
+                                              (size_t)0);
         if(total_vsizes > vals.size) return Status::InvalidArg;
 
         MDB_txn* txn = nullptr;
@@ -367,7 +367,7 @@ class LMDBDatabase : public DocumentStoreMixin<DatabaseInterface> {
         size_t key_offset = 0;
         size_t total_ksizes = std::accumulate(ksizes.data,
                                               ksizes.data + ksizes.size,
-                                              0);
+                                              (size_t)0);
         if(total_ksizes > keys.size) return Status::InvalidArg;
 
         MDB_txn* txn = nullptr;
