+++ src/backends/set.cpp
@@ -263,12 +263,12 @@ class SetDatabase : public DatabaseInterface {
 
         size_t total_ksizes = std::accumulate(ksizes.data,
                                               ksizes.data + ksizes.size,
-                                              0);
+                                              (size_t)0);
         if(total_ksizes > keys.size) return Status::InvalidArg;
 
         size_t total_vsizes = std::accumulate(vsizes.data,
                                               vsizes.data + vsizes.size,
-                                              0);
+                                              (size_t)0);
         if(total_vsizes != 0) return Status::InvalidArg;
 
         if(mode_exist_only) {
