+++ src/backends/unordered_set.cpp
@@ -209,12 +209,12 @@ class UnorderedSetDatabase : public DatabaseInterface {
 
         size_t total_ksizes = std::accumulate(ksizes.data,
                                               ksizes.data + ksizes.size,
-                                              0);
+                                              (size_t)0);
         if(total_ksizes > keys.size) return Status::InvalidArg;
 
         size_t total_vsizes = std::accumulate(vsizes.data,
                                               vsizes.data + vsizes.size,
-                                              0);
+                                              (size_t)0);
         if(total_vsizes > 0) return Status::InvalidArg;
 
         if(mode & YOKAN_MODE_EXIST_ONLY) {
